Restructure the codebase
Created different file for each struct. Separated APIResult and Attributes struct from Tutorial struct file.
