+++ Practical Example/RW Take Home/RW Take Home/Model/APIResult.swift
@@ -1,0 +1,13 @@
+//
+//  APIResult.swift
+//  RW Take Home
+//
+//  Created by Zoha on 8/1/20.
+//  Copyright © 2020 Zoha. All rights reserved.
+//
+
+import Foundation
+
+struct APIResult: Codable {
+  let data: [Tutorial]
+}
